split plotting file into plot submodule
