+++ dabl/__init__.py
@@ -1,6 +1,6 @@
 from .preprocessing import EasyPreprocessor, clean, detect_types
 from .models import SimpleClassifier
-from .plotting import plot_supervised
+from .plot.supervised import plot_supervised
 from .explain import explain
 
 __all__ = ['EasyPreprocessor', 'SimpleClassifier', 'explain', 'clean',
