+++ dabl/plot/__init__.py
@@ -1,0 +1,9 @@
+from .supervised import (plot_supervised, plot_classification_categorical,
+                         plot_regression_categorical,
+                         plot_classification_continuous,
+                         plot_regression_continuous)
+
+__all__ = [
+    'find_pretty_grid', 'plot_supervised', 'plot_classification_categorical',
+    'plot_classification_continuous', 'plot_regression_categorical',
+    'plot_regression_continuous']
