+++ dabl/search.py
@@ -1,6 +1,5 @@
 from math import ceil, floor, log
 from abc import abstractmethod
-from copy import deepcopy
 
 import numpy as np
 from sklearn.model_selection._search import _check_param_grid
@@ -58,12 +57,12 @@ class BaseSuccessiveHalving(CustomBaseSearchCV):
 
     def _check_input_parameters(self, X, y, groups):
 
-        if (self.budget_on != 'n_samples' and
-                self.budget_on not in self.estimator.get_params()):
+        if (self.budget_on != 'n_samples'
+                and self.budget_on not in self.estimator.get_params()):
             raise ValueError(
                 'Cannot budget on parameter {} which is not supported '
                 'by estimator {}'.format(self.budget_on,
-                                        self.estimator.__class__.__name__))
+                                         self.estimator.__class__.__name__))
 
         if isinstance(self.max_budget, str) and self.max_budget != 'auto':
             raise ValueError(
@@ -211,7 +210,8 @@ class BaseSuccessiveHalving(CustomBaseSearchCV):
                 # train_test_split because it complains about testset being too
                 # small in some cases
                 indexes = rng.choice(X.shape[0], r_i)
-                X_iter, y_iter = safe_indexing(X, indexes), safe_indexing(y, indexes)
+                X_iter = safe_indexing(X, indexes)
+                y_iter = safe_indexing(y, indexes)
             else:
                 # Need copy so that r_i of next iteration do not overwrite
                 candidate_params = [c.copy() for c in candidate_params]
